Update README.md
Create a JavaFX GUI to convert temperature reported in any of the units, Kelvin, Fahrenheit, and Celsius from one unit to the other by mimicking how Google does (Links to an external site.) it. Note that you only need to implement the converter for temperature only though in Google GUI, they are doing it for bunch of other quantities.